# docs/swagger.yaml
basePath: /api
definitions:
...
  model.Organisation:
    properties:
      createdAt:
        type: string
      deletedAt:
        type: string
      id:
        type: integer
      name:
        type: string
      updatedAt:
        type: string
    type: object
...
  model.Qualification:
    properties:
      createdAt:
        type: string
      deletedAt:
        type: string
      id:
        type: integer
      name:
        type: string
      organisation:
        $ref: '#/definitions/model.Organisation'
      organisationId:
        type: integer
      updatedAt:
        type: string
    type: object
...
  model.Slide:
    properties:
      createdAt:
        type: string
      deletedAt:
        type: string
      description:
        type: string
      hash:
        type: string
      id:
        type: integer
      name:
        type: string
      syllabus:
        $ref: '#/definitions/model.Syllabus'
      syllabusId:
        type: integer
      updatedAt:
        type: string
    type: object
  model.SlideCreateEditRequest:
    properties:
      description:
        type: string
      id:
        type: integer
      name:
        type: string
      syllabusId:
        type: integer
    type: object
  model.SlideQueryRequest:
    properties:
      id:
        type: integer
      pageIndex:
        type: integer
      pageSize:
        type: integer
      syllabusId:
        type: integer
    type: object
  model.Syllabus:
    properties:
      code:
        type: string
      createdAt:
        type: string
      deletedAt:
        type: string
      id:
        type: integer
      name:
        type: string
      qualification:
        $ref: '#/definitions/model.Qualification'
      qualificationId:
        type: integer
      updatedAt:
        type: string
    type: object
...
info:
  contact:
    name: LTEdu Support
  description: LTEdu 在线教育平台后端接口文档
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: LTEdu API
  version: "1.0"
paths:
...
  /v1/slide/all:
    post:
      consumes:
      - application/json
      description: 获取全部课件列表（不分页）
      parameters:
      - description: 查询条件
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/model.SlideQueryRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            additionalProperties: true
            type: object
        "400":
          description: 参数错误
          schema:
            additionalProperties: true
            type: object
      summary: 获取所有课件
      tags:
      - 课程
  /v1/slide/create:
    post:
      consumes:
      - application/json
      description: 创建新课件
      parameters:
      - description: 课件信息
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/model.SlideCreateEditRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            additionalProperties: true
            type: object
        "400":
          description: 参数错误
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: 创建课件
      tags:
      - 课程
  /v1/slide/delete:
    post:
      consumes:
      - application/json
      description: 删除指定课件
      parameters:
      - description: 课件ID
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/model.Slide'
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            additionalProperties: true
            type: object
        "400":
          description: 参数错误
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: 删除课件
      tags:
      - 课程
  /v1/slide/edit:
    post:
      consumes:
      - application/json
      description: 修改课件信息
      parameters:
      - description: 课件信息
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/model.SlideCreateEditRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            additionalProperties: true
            type: object
        "400":
          description: 参数错误
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: 编辑课件
      tags:
      - 课程
  /v1/slide/getById:
    post:
      consumes:
      - application/json
      description: 根据课件ID获取课件详情
      parameters:
      - description: 课件ID
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/model.SlideQueryRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            additionalProperties: true
            type: object
        "400":
          description: 参数错误
          schema:
            additionalProperties: true
            type: object
      summary: 根据ID获取课件
      tags:
      - 课程
  /v1/slide/link:
    post:
      consumes:
      - application/json
      description: 获取课件的访问哈希值
      parameters:
      - description: 课件ID
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/model.SlideQueryRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            additionalProperties: true
            type: object
        "400":
          description: 参数错误
          schema:
            additionalProperties: true
            type: object
      summary: 获取课件链接
      tags:
      - 课程
  /v1/slide/list:
    post:
      consumes:
      - application/json
      description: 分页查询课件列表
      parameters:
      - description: 查询条件
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/model.SlideQueryRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            additionalProperties: true
            type: object
        "400":
          description: 参数错误
          schema:
            additionalProperties: true
            type: object
      summary: 获取课件列表
      tags:
      - 课程
  /v1/statistic/orderCreated:
    get:
      consumes:
      - application/json
      description: 获取指定时间范围内的订单创建数量统计
      parameters:
      - description: 开始日期 (2006-01-02)
        in: query
        name: startAt
        type: string
      - description: 结束日期 (2006-01-02)
        in: query
        name: endAt
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: 订单创建统计
      tags:
      - 系统设置
  /v1/statistic/orderPaidCount:
    get:
      consumes:
      - application/json
      description: 获取指定时间范围内的订单支付数量统计
      parameters:
      - description: 开始日期 (2006-01-02)
        in: query
        name: startAt
        type: string
      - description: 结束日期 (2006-01-02)
        in: query
        name: endAt
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: 订单支付数量统计
      tags:
      - 系统设置
  /v1/statistic/orderPaidSum:
    get:
      consumes:
      - application/json
      description: 获取指定时间范围内的订单支付金额统计
      parameters:
      - description: 开始日期 (2006-01-02)
        in: query
        name: startAt
        type: string
      - description: 结束日期 (2006-01-02)
        in: query
        name: endAt
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: 订单支付金额统计
      tags:
      - 系统设置
  /v1/statistic/userRegister:
    get:
      consumes:
      - application/json
      description: 获取指定时间范围内的用户注册数量统计
      parameters:
      - description: 开始日期 (2006-01-02)
        in: query
        name: startAt
        type: string
      - description: 结束日期 (2006-01-02)
        in: query
        name: endAt
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: 用户注册统计
      tags:
      - 系统设置
...
securityDefinitions:
  BearerAuth:
    description: '输入 "Bearer {token}" 格式的 JWT，例如: Bearer eyJhbGci...'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"